# docs/swagger.yaml
definitions:
  main.CreatePaymentPayload:
    properties:
      seat_id:
        type: integer
      session_id:
        type: integer
    required:
    - seat_id
    - session_id
    type: object
  main.CreateRoomPayload:
    properties:
      capacity:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - capacity
    - name
    type: object
...
  main.UpdateRoomPayload:
    properties:
      capacity:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
  payment.PaymentResponse:
    properties:
      order_id:
        type: integer
      status:
        type: string
      url:
        type: string
    type: object
...
  store.Room:
    properties:
      capacity:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Tikceto
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tikceto
paths:
...
  /payments/create:
    post:
      consumes:
      - application/json
      description: Creates a payment link for a ticket through LiqPay
      parameters:
      - description: Payment request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreatePaymentPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payment.PaymentResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a payment link
      tags:
      - payments
  /payments/status/{orderID}:
    put:
      consumes:
      - application/json
      description: Validates the payment status by order ID
      parameters:
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Validate payment status
      tags:
      - payments
  /rooms:
    post:
      consumes:
      - application/json
      description: Creates a room
      parameters:
      - description: Room payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateRoomPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Room'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a room
      tags:
      - rooms
  /rooms/:
    get:
      consumes:
      - application/json
      description: Fetches all rooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Room'
            type: array
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Fetches all rooms
      tags:
      - rooms
  /rooms/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Deletes a room
      tags:
      - rooms
    get:
      consumes:
      - application/json
      description: Fetches a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Room'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Fetches a room
      tags:
      - rooms
    patch:
      consumes:
      - application/json
      description: Updates a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Room payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateRoomPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Room'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Updates a room
      tags:
      - rooms
...
securityDefinitions:
  ApiKeyAuth:
    description: Use the Authorization header with a Bearer token to authenticate
      API requests.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"